+++ Content/Core/Items/Accessories/Combat/Magic/Manamaster.cs
@@ -14,7 +14,7 @@ namespace TLR.Content.Core.Items.Accessories.Combat.Magic
 		{
 			Item.width = 28;
 			Item.height = 32;
-			Item.value = Item.sellPrice(gold: 20);
+			Item.value = Item.sellPrice(gold: 15);
 			Item.rare = ItemRarityID.Lime;
 			Item.accessory = true;
 			Item.ResearchUnlockCount = 1;
@@ -25,7 +25,7 @@ namespace TLR.Content.Core.Items.Accessories.Combat.Magic
             player.manaFlower = true;
             player.manaCost -= 0.12f;
             player.aggro -= 400;
-			player.statManaMax2 += 20;
+			player.statManaMax2 += 40;
             player.GetDamage(DamageClass.Magic) += 0.15f;
             player.magicCuffs = true;
             player.manaMagnet = true;