+++ Content/Core/Items/Accessories/Combat/Melee/HallowedGauntlet.cs
@@ -16,7 +16,7 @@ namespace TLR.Content.Core.Items.Accessories.Combat.Melee
 		{
 			Item.width = 22;
 			Item.height = 28;
-			Item.value = Item.sellPrice(gold: 30);
+			Item.value = Item.sellPrice(gold: 15);
 			Item.rare = ItemRarityID.Yellow;
 			Item.accessory = true;
 			Item.ResearchUnlockCount = 1;
@@ -39,25 +39,25 @@ namespace TLR.Content.Core.Items.Accessories.Combat.Melee
 
         public override void AddRecipes()
 		{
-			Recipe recipe = CreateRecipe();
-			recipe.AddIngredient(ItemID.FireGauntlet);
-			recipe.AddIngredient(ItemID.YoyoBag);
-            recipe.AddIngredient(ItemID.FleshKnuckles);
-			recipe.AddIngredient(ItemID.SharpeningStation);
-			recipe.AddIngredient(ItemID.HallowedBar, 15);
-			recipe.AddTile(TileID.MythrilAnvil);
-            recipe.AddTile(TileID.TinkerersWorkbench);
-			recipe.Register();
-			Recipe recipe2 = CreateRecipe();
-			recipe2.AddIngredient(ItemID.BerserkerGlove);
-			recipe2.AddIngredient(ItemID.AvengerEmblem);
-			recipe2.AddIngredient(ItemID.MagmaStone);
-			recipe2.AddIngredient(ItemID.YoyoBag);
-			recipe2.AddIngredient(ItemID.SharpeningStation);
-			recipe2.AddIngredient(ItemID.HallowedBar, 15);
-			recipe2.AddTile(TileID.MythrilAnvil);
-            recipe2.AddTile(TileID.TinkerersWorkbench);
-			recipe2.Register();
+			CreateRecipe()
+				.AddIngredient(ItemID.FireGauntlet)
+				.AddIngredient(ItemID.YoyoBag)
+				.AddIngredient(ItemID.FleshKnuckles)
+				.AddIngredient(ItemID.SharpeningStation)
+				.AddIngredient(ItemID.HallowedBar, 18)
+				.AddTile(TileID.TinkerersWorkbench)
+				.AddTile(TileID.MythrilAnvil)
+				.Register();
+			CreateRecipe()
+				.AddIngredient(ItemID.BerserkerGlove)
+				.AddIngredient(ItemID.YoyoBag)
+				.AddIngredient(ItemID.AvengerEmblem)
+				.AddIngredient(ItemID.MagmaStone)
+				.AddIngredient(ItemID.SharpeningStation)
+				.AddIngredient(ItemID.HallowedBar, 18)
+				.AddTile(TileID.TinkerersWorkbench)
+				.AddTile(TileID.MythrilAnvil)
+				.Register();
 		}
 	}
 }