+++ Content/Core/Items/Accessories/Movement/Boots/MechaBoots.cs
@@ -7,14 +7,13 @@ namespace TLR.Content.Core.Items.Accessories.Movement.Boots
 	// This is a basic item template.
 	// Please see tModLoader's ExampleMod for every other example:
 	// https://github.com/tModLoader/tModLoader/tree/stable/ExampleMod
-    [AutoloadEquip(EquipType.Shoes, EquipType.Wings)]
+    [AutoloadEquip(EquipType.Shoes)]
 	public class MechaBoots : ModItem
 	{
         // The Display Name and Tooltip of this item can be edited in the 'Localization/en-US_Mods.TLR.hjson' file.
         public override bool IsLoadingEnabled(Mod mod) => ModContent.GetInstance<TLRConfigServer>().Spriteless >= 1;
         public override void SetDefaults()
 		{
-            if (ModContent.GetInstance<TLRConfigServer>().WingedBoots) { Item.CloneDefaults(ItemID.Jetpack); }
 			Item.width = 32;
 			Item.height = 32;
 			Item.value = Item.sellPrice(gold: 20);
@@ -26,7 +25,7 @@ namespace TLR.Content.Core.Items.Accessories.Movement.Boots
         public override void UpdateAccessory(Player player, bool hideVisual)
         {
             player.accRunSpeed = 7f;
-            player.moveSpeed += 0.16f;
+            player.moveSpeed += 0.12f;
             player.vanityRocketBoots = player.rocketBoots = 4;
             player.iceSkate = true;
             player.waterWalk = true;
@@ -36,18 +35,21 @@ namespace TLR.Content.Core.Items.Accessories.Movement.Boots
 			player.accFlipper = true;
 			player.accDivingHelm = true;
 			player.arcticDivingGear = true;
-			player.endurance += 0.04f;
+			player.frogLegJumpBoost = true;
+			player.noFallDmg = true;
+			player.autoJump = true;
         }
 
         public override void AddRecipes()
 		{
-			Recipe recipe = CreateRecipe();
-			recipe.AddIngredient(ItemID.TerrasparkBoots);
-            if (ModContent.GetInstance<TLRConfigServer>().WingedBoots) { recipe.AddIngredient(ItemID.Jetpack); }
-			recipe.AddIngredient(ItemID.SoulofFright, 10);
-            recipe.AddIngredient(ItemID.SoulofFlight, 20);
-			recipe.AddTile(TileID.TinkerersWorkbench);
-			recipe.Register();
+			CreateRecipe()
+				.AddIngredient(ItemID.TerrasparkBoots)
+				.AddIngredient(ItemID.AmphibianBoots)
+				.AddIngredient(ItemID.SoulofMight, 10)
+				.AddIngredient(ItemID.SoulofFright, 10)
+				.AddIngredient(ItemID.SoulofMight, 10)
+				.AddTile(TileID.TinkerersWorkbench)
+				.Register();
 		}
 	}
 }