+++ Content/Core/Items/Accessories/Movement/Boots/Shroomts.cs
@@ -7,14 +7,13 @@ namespace TLR.Content.Core.Items.Accessories.Movement.Boots
 	// This is a basic item template.
 	// Please see tModLoader's ExampleMod for every other example:
 	// https://github.com/tModLoader/tModLoader/tree/stable/ExampleMod
-    [AutoloadEquip(EquipType.Shoes, EquipType.Wings)]
+    [AutoloadEquip(EquipType.Shoes)]
 	public class Shroomts : ModItem
 	{
         // The Display Name and Tooltip of this item can be edited in the 'Localization/en-US_Mods.TLR.hjson' file.
         public override bool IsLoadingEnabled(Mod mod) => ModContent.GetInstance<TLRConfigServer>().Spriteless >= 1;
         public override void SetDefaults()
 		{
-            if (ModContent.GetInstance<TLRConfigServer>().WingedBoots) { Item.CloneDefaults(ItemID.Hoverboard); }
 			Item.width = 32;
 			Item.height = 28;
 			Item.value = Item.sellPrice(gold: 25);
@@ -26,30 +25,32 @@ namespace TLR.Content.Core.Items.Accessories.Movement.Boots
         public override void UpdateAccessory(Player player, bool hideVisual)
         {
             player.accRunSpeed = 8f;
-            player.moveSpeed += 0.33f;
+            player.moveSpeed += 0.16f;
             player.vanityRocketBoots = player.rocketBoots = 5;
             player.iceSkate = true;
             player.waterWalk = true;
             player.waterWalk2 = true;
             player.fireWalk = true;
-            player.lavaMax += 900;
+            player.lavaMax += 600;
 			player.accFlipper = true;
 			player.accDivingHelm = true;
 			player.arcticDivingGear = true;
             player.dashType = 1;
             player.spikedBoots = 2;
             player.blackBelt = true;
-            player.endurance += 0.06f;
+			player.frogLegJumpBoost = true;
+			player.noFallDmg = true;
+			player.autoJump = true;
         }
 
         public override void AddRecipes()
 		{
-			Recipe recipe = CreateRecipe();
-			recipe.AddIngredient(ModContent.ItemType<MechaBoots>());
-            if (ModContent.GetInstance<TLRConfigServer>().WingedBoots) { recipe.AddIngredient(ItemID.Hoverboard); }
-			recipe.AddIngredient(ItemID.MasterNinjaGear);
-			recipe.AddTile(TileID.TinkerersWorkbench);
-			recipe.Register();
+			CreateRecipe()
+				.AddIngredient(ModContent.ItemType<MechaBoots>())
+				.AddIngredient(ItemID.MasterNinjaGear)
+				.AddIngredient(ItemID.ShroomiteBar, 18)
+				.AddTile(TileID.TinkerersWorkbench)
+				.Register();
 		}
 	}
 }